requirements.txt
# Core HTML to Markdown conversion
html2text>=2020.1.16
beautifulsoup4>=4.12.0
lxml>=4.9.0

# YAML processing for configuration
PyYAML>=6.0.1

# Path handling and file operations
pathlib>=1.0.1

# Handlebars template processing
pybars3>=0.9.7



